Rename BrevModulContext -> useBrevModul og flytt til passende mappe
