+++ src/frontend/komponenter/Felleskomponenter/Hendelsesoversikt/BrevModul/useBrevModul.ts
@@ -4,29 +4,29 @@ import type { Avhengigheter, FeltState } from '@navikt/familie-skjema';
 import { feil, ok, useFelt, useSkjema, Valideringsstatus } from '@navikt/familie-skjema';
 import { hentDataFraRessurs } from '@navikt/familie-typer';
 
-import { useBehandling } from './behandlingContext/BehandlingContext';
-import { useFagsakContext } from './Fagsak/FagsakContext';
-import type { ISelectOptionMedBrevtekst } from '../komponenter/Felleskomponenter/Hendelsesoversikt/BrevModul/typer';
-import { Brevmal } from '../komponenter/Felleskomponenter/Hendelsesoversikt/BrevModul/typer';
-import type { IBehandling } from '../typer/behandling';
-import { BehandlingKategori } from '../typer/behandlingstema';
-import type { IManueltBrevRequestPåBehandling } from '../typer/dokument';
-import { FagsakType } from '../typer/fagsak';
-import type { IGrunnlagPerson } from '../typer/person';
-import { PersonType } from '../typer/person';
-import type { IBarnMedOpplysninger, Målform } from '../typer/søknad';
+import type { ISelectOptionMedBrevtekst } from './typer';
+import { Brevmal } from './typer';
+import { useBehandling } from '../../../../context/behandlingContext/BehandlingContext';
+import { useFagsakContext } from '../../../../context/Fagsak/FagsakContext';
+import type { IBehandling } from '../../../../typer/behandling';
+import { BehandlingKategori } from '../../../../typer/behandlingstema';
+import type { IManueltBrevRequestPåBehandling } from '../../../../typer/dokument';
+import { FagsakType } from '../../../../typer/fagsak';
+import type { IGrunnlagPerson } from '../../../../typer/person';
+import { PersonType } from '../../../../typer/person';
+import type { IBarnMedOpplysninger, Målform } from '../../../../typer/søknad';
 import {
     hentMuligeBrevmalerImplementering,
     mottakersMålformImplementering,
-} from '../utils/brevmal';
-import type { IsoDatoString } from '../utils/dato';
-import { dateTilIsoDatoStringEllerUndefined, validerGyldigDato } from '../utils/dato';
-import { useDeltBostedFelter } from '../utils/deltBostedSkjemaFelter';
-import type { IFritekstFelt } from '../utils/fritekstfelter';
+} from '../../../../utils/brevmal';
+import type { IsoDatoString } from '../../../../utils/dato';
+import { dateTilIsoDatoStringEllerUndefined, validerGyldigDato } from '../../../../utils/dato';
+import { useDeltBostedFelter } from '../../../../utils/deltBostedSkjemaFelter';
+import type { IFritekstFelt } from '../../../../utils/fritekstfelter';
 import {
     genererIdBasertPåAndreFritekstKulepunkter,
     lagInitiellFritekst,
-} from '../utils/fritekstfelter';
+} from '../../../../utils/fritekstfelter';
 
 export const useBrevModul = () => {
     const { behandling } = useBehandling();
